# Remove dynamic calls