--- a/analysis_options.yaml
+++ b/analysis_options.yaml
@@ -2,6 +2,7 @@ linter:
   rules:
   - always_require_non_null_named_parameters
   - annotate_overrides
+  - avoid_dynamic_calls
   - avoid_empty_else
   - avoid_function_literals_in_foreach_calls
   - avoid_init_to_null