--- a/protobuf/analysis_options.yaml
+++ b/protobuf/analysis_options.yaml
@@ -2,6 +2,7 @@ include: package:lints/recommended.yaml
 
 linter:
   rules:
+    - avoid_dynamic_calls
     - comment_references
     - directives_ordering
     - prefer_relative_imports