--- a/protobuf/test/message_test.dart
+++ b/protobuf/test/message_test.dart
@@ -22,6 +22,9 @@ class Rec extends MockMessage {
 Matcher throwsError(Type expectedType, String expectedMessage) =>
     throwsA(predicate((dynamic x) {
       expect(x.runtimeType, expectedType);
+      // TODO(https://github.com/google/protobuf.dart/issues/578): Remove
+      // remaining dynamic calls.
+      // ignore: avoid_dynamic_calls
       expect(x!.message, expectedMessage);
       return true;
     }));