Work towards https://github.com/google/protobuf.dart/issues/578

Many of these fixes were guesses. Function `f` takes a `dynamic` which is later indexed with `[]` in a sub-path of the function. Function `g` passes a value to function `f` but that value is also `dynamic`. `dynamic` all the way down. The bright side is that these guesses are less guess-y than dynamic invocations. :)

I did not land any fixes in tests. This PR is to improve app size.