@mraleph 2022-02-10
> OK I've removed all explicit casts. I think the owners of this package should be making the tougher calls here.

Lets sit on this PR and return to it in a bit (~1 month). I don't have time to run benchmarks right now, but in a month we will have somebody on board who can take care of this.