@osa1 2022-03-17
I compared the benchmark file sizes and runtime numbers.

I haven't checked if if the benchmarks exercise the code paths changed with this PR.

## Runtime perf on the VM

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 5,471 μs  | 5,331 μs  | -140 μs, -2.55%
| ToBinary(RunTime)   | 6,302 μs  | 6,277 μs  | -25 μs, -0.39%
| ToJson(RunTime)     | 24,879 μs | 24,658 μs | -221 μs, -0.88%
| FromJson(RunTime)   | 20,199 μs | 20,277 μs | +78 μs, +0.38%

## Runtime perf when compiled to JS, with default parameters

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 31,952 μs | 31,841 μs | -111 μs, -0.34%
| ToBinary(RunTime)   | 24,451 μs | 24,646 μs | +195 μs, +0.79%
| ToJson(RunTime)     | 50,000 μs | 49,707 μs | -293 μs, -0.58%
| FromJson(RunTime)   | 34,101 μs | 34,271 μs | +170 μs, +0.49%

## Benchmark JS sizes, with default parameters

| master   | #579     | Difference
| -------- | -------- | -----------
| 981,993 characters  | 982,132 characters | +139 characters, +0.01%

## Runtime perf when compiled to JS, with -O3

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 22,505 μs | 22,400 μs | -105 μs, -0.46%
| ToBinary(RunTime)   | 19,821 μs | 19,355 μs | -466 μs, -2.35%
| ToJson(RunTime)     | 36,381 μs | 36,381 μs | 0 μs
| FromJson(RunTime)   | 25,468 μs | 25,626 μs | +158 μs, +0.62%

## Benchmark JS sizes, with -O3

| master   | #579     | Difference
| -------- | -------- | -----------
| 335,815 characters  | 335,836 characters | +21 characters, +0.006%

---

> I think the owners of this package should be making the tougher calls here

Assuming the benchmarks exercise the code paths changed in this PR, results look OK to me. What calls remain to be made here?