@osa1 2022-03-17
In my comment above I compared master with the last commit in this branch, which reverted some of the dynamic call fixes for the reasons explained in https://github.com/google/protobuf.dart/pull/579#pullrequestreview-879476091.

If I revert the last commit and bring back casts with `as` syntax, this is the results I get in JS benchmarks:

## Runtime perf when compiled to JS, with default parameters

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 31,000 μs | 31,453 μs | +453 μs, +1.46%
| ToBinary(RunTime)   | 23,279 μs | 24,512 μs | +1,233 μs, +5.29%
| ToJson(RunTime)     | 48,071 μs | 48,804 μs | +1,524 μs, +1.52%
| FromJson(RunTime)   | 34,101 μs | 33,383 μs | -718 μs, -2.10%

## Benchmark JS sizes, with default parameters

| master   | #579     | Difference
| -------- | -------- | -----------
| 981,993 characters  | 978,553 characters | -3,440 characters, -0.35%

## Runtime perf when compiled to JS, with -O3

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 22,573 μs | 22,651 μs | +78 μs, +0.34%
| ToBinary(RunTime)   | 19,403 μs | 19,009 μs | -394 μs, -2.93%
| ToJson(RunTime)     | 36,303 μs | 38,377 μs | +2,074 μs, +5.71%
| FromJson(RunTime)   | 25,481 μs | 25,679 μs | +198 μs, +0.77%

## Benchmark JS sizes, with -O3

| master   | #579     | Difference
| -------- | -------- | -----------
| 335,815 characters  | 335,571 characters | -244 characters, -0.07%

---

It seems like `as` adds some overhead in 3 out of 4 benchmarks with default optimizations. I will now try to replace some of the explicit casts with implicit casts and see if it makes any difference when compiled to JS with `-O3`.