@osa1 2022-03-17
With the explicit casts replaced with implicit casts, compiled with -O3:

## Runtime perf when compiled to JS, with -O3

| Benchmark           | master   | #579     | Difference
| ------------------- | -------- | -------- | -----------
| FromBinary(RunTime) | 22,573 μs | 22,617 μs | +70 μs, +0.31%
| ToBinary(RunTime)   | 19,403 μs | 19,230 μs | -173 μs, -0.89%
| ToJson(RunTime)     | 36,303 μs | 37,905 μs | +1,602 μs, +4.41%
| FromJson(RunTime)   | 25,481 μs | 25,455 μs | -26 μs, -0.10%

## Benchmark JS sizes, with -O3

| master   | #579     | Difference
| -------- | -------- | -----------
| 335,815 characters  | 335,402 characters | -413 characters, -0.12%

This is much better, but one of the benchmarks still regresses 4.41% (with explicit casts: 5.71%).