@osa1 2022-06-10
After realizing how costly type casts are when targeting JS (see my comment above), I re-implemented this PR myself in https://github.com/osa1/protobuf.dart/commit/1dcf951622893c0d8a8074643a99a58c1c8cdb05.

Results: (copied from commit message)

## File size:

|     | Before          | After           | Diff                |
|-----|-----------------|-----------------|---------------------|
| JS  | 348,393 bytes   | 344,804 bytes   | -3,589 bytes -1.0%  |
| AOT | 7,235,848 bytes | 7,206,256 bytes | -29,592 bytes -0.4% |

## JS runtime:

|                      | Before    | After     | Diff              |
|----------------------|-----------|-----------|-------------------|
| FromBinary           | 43,361 us | 39,588 us | -3,773 us, -8.7%  |
| ToBinary             | 37,796 us | 36,636 us | -1,160 us, -3.0%  |
| FromJson             | 47,465 us | 43,085 us | -4,380 us, -9.2%  |
| ToJson               | 75,185 us | 70,689 us | -4,496 us, -5.9%  |
| FromProto3JsonString | 37,555 us | 34,500 us | -3,055 us, -8.1%  |
| ToProto3JsonString   | 76,259 us | 70,068 us | -6,191 us, -8.1%  |
| FromProto3JsonObject | 17,213 us | 17,025 us | -188 us, -1.0%    |
| ToProto3JsonObject   | 21,836 us | 19,607 us | -2,229 us, -10.2% |
| HashCode             | 12,151 us | 12,029 us | -122 us, -1.0%    |

## AOT runtime:

|                      | Before    | After     | Diff              |
|----------------------|-----------|-----------|-------------------|
| FromBinary           | 6,428 us  | 6,581 us  | +153 us, +2.3%    |
| ToBinary             | 12,560 us | 12,893 us | +333 us, +2.6%    |
| FromJson             | 27,877 us | 28,274 us | +397 us, +1.4%    |
| ToJson               | 34,721 us | 36,316 us | +1595 us, +4.5%   |
| FromProto3JsonString | 31,825 us | 32,587 us | +752 us, +2.3%    |
| ToProto3JsonString   | 40,089 us | 39,135 us | -954 us, -2.3%    |
| FromProto3JsonObject | 9,002 us  | 9,091 us  | +91 us, +0.98%    |
| ToProto3JsonObject   | 11,163 us | 10,925 us | -238 us, -2.1%    |
| HashCode             | 5,108 us  | 5,007 us  | -101 us, -1.9%    |

---

Avoiding dynamic calls and explicit type casts is a big win on JS.

However on native it's slower on most benchmarks.

It's unclear to me whether we want to make that trade.

To me the fundamental problem is that JS and native targets are wildly different, and what is an optimization for one can easily be a pessimization in other. The only solution that I can see is conditional compilation, which is difficult to do in Dart, and too much conditional compilation also makes testing more difficult and code unmaintainable.

I'm just hoping that Wasm will be mature enough soon (the spec, implementations in browsers, dart2wasm) and we will be able to target Wasm instead of JS..